refactored youtube poller structure
