+++ src/helpers/fetchData.js
@@ -1,5 +1,7 @@
 /* eslint-disable no-console */
 const fetch = require('node-fetch');
+const { red, green } = require('colors/safe');
+const Video = require('../api/video/video.model');
 const { YOUTUBE_API_KEY, YOUTUBE_CHANNEL_ID } = require('../config');
 
 async function fetchLatestYoutubeVideos({ maxResults, publishedAfter }) {
@@ -37,4 +39,28 @@ async function fetchLatestYoutubeVideos({ maxResults, publishedAfter }) {
   }
 }
 
-module.exports = { fetchLatestYoutubeVideos };
+async function fetchVideosJob() {
+  try {
+    const video = await Video.findOne({}).sort({ date: -1 });
+    // Check if db has at least one video
+    if (video) {
+      // Transforms date format to the Youtube-API standard.
+      const lastDate = video.date.toISOString();
+
+      const fetchedVideos = await fetchLatestYoutubeVideos({ publishedAfter: lastDate });
+
+      if (fetchedVideos.length > 0) {
+        fetchedVideos.forEach(async newVideo => {
+          if (newVideo.date !== lastDate) {
+            const { name } = await new Video(newVideo).save();
+            console.log(green(`[cron-job] Added new video from Youtube: ${name}, at ${new Date().toISOString()}`));
+          }
+        });
+      }
+    }
+  } catch (err) {
+    console.error(red(`[cron-job-error] ${err}`));
+  }
+}
+
+module.exports = { fetchLatestYoutubeVideos, fetchVideosJob };
